Added Update and GetByID on Accounts

+++ SimpleTradingApp.Application/Services/AccountsService.cs
@@ -24,6 +24,39 @@ public class AccountsService : IAccountsService
 
     public async Task<bool> DeleteAccount(Guid accountId)
     {
+        var account = await _accountsRepository.GetAccountById(accountId);
+
+        if (account == null)
+        {
+            return false;
+        }
+
         return await _accountsRepository.DeleteAccount(accountId);
+    }
+
+    public async Task<AccountResponse?> GetAccountById(Guid accountId)
+    {
+        var account = await _accountsRepository.GetAccountById(accountId, true);
+
+        if (account == null)
+        {
+            return null;
+        }
+
+        return account.ToAccountResponse();
+    }
+
+    public async Task<AccountResponse?> UpdateAccount(UpdateAccountDto updateAccountDto)
+    {
+        var account = await _accountsRepository.GetAccountById(updateAccountDto.ID);
+
+        if (account == null)
+        {
+            return null;
+        }
+
+        var updatedAccount = await _accountsRepository.UpdateAccount(updateAccountDto.ToAccount());
+
+        return updatedAccount?.ToAccountResponse();
     }
 }
